extract parsing module
